Separated payload models in contracts Fixed Automapper registration Fixed Automapper registration in tests
