+++ src/DataIngestor.Contracts/EnergyReadingPayload.cs
@@ -1,0 +1,6 @@
+using System.Text.Json.Serialization;
+
+namespace DataIngestor.Contracts;
+
+public record EnergyReadingPayload(
+    [property: JsonPropertyName("energy")] double Amount) : MetricReadingPayload;
