+++ src/DataIngestor.Contracts/MotionReadingPayload.cs
@@ -1,0 +1,6 @@
+using System.Text.Json.Serialization;
+
+namespace DataIngestor.Contracts;
+
+public record MotionReadingPayload(
+    [property: JsonPropertyName("motionDetected")] bool IsDetected) : MetricReadingPayload;
